Moved all SEP-10 related functions from Client to SEP10 class
